+++ lib/plausible_web/views/layout_view.ex
@@ -40,6 +40,27 @@ defmodule PlausibleWeb.LayoutView do
     end
   end
 
+  defmodule JWT do
+    use Joken.Config
+  end
+
+  def feedback_link(user) do
+    token_params = %{
+      "id" => user.id,
+      "email" => user.email,
+      "name" => user.name,
+      "imageUrl" => Plausible.Auth.User.profile_img_url(user)
+    }
+
+    case JWT.generate_and_sign(token_params) do
+      {:ok, token, _claims} ->
+        "https://feedback.plausible.io/sso/#{token}?returnUrl=https://feedback.plausible.io"
+
+      _ ->
+        "https://feedback.plausible.io"
+    end
+  end
+
   def home_dest(conn) do
     if conn.assigns[:current_user] do
       "/sites"